requirements.txt
 fastapi==0.115.0
 uvicorn[standard]==0.30.6
 SQLAlchemy==2.0.36
 psycopg2-binary==2.9.9
 pydantic==2.9.0
 python-jose[cryptography]==3.3.0
 passlib[bcrypt]==1.7.4
 boto3==1.35.10
 langchain==0.3.4
 langchain-community==0.3.3
 pgvector==0.3.2
 numpy==1.26.4
 pdfplumber==0.11.4
 PyMuPDF==1.24.9
 tiktoken==0.7.0
 tenacity==8.3.0
 httpx==0.27.2
 redis==5.1.1
 python-dotenv==1.0.1
# AI CORE RAG API Service - Essential Dependencies

# FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.27.1

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for Qdrant integration
httpx==0.28.1

# OpenAI integration
openai==1.35.0

# LangChain for RAG pipeline
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
